# Back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "name": "Soporte",
            "email": "soporte@terciariourquiza.edu.ar"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/horarios/grados": {
            "get": {
                "tags": [
                    "Grado"
                ],
                "summary": "Obtener todos los grados",
                "description": "Devuelve todos los grados",
                "operationId": "getGrados",
                "responses": {
                    "200": {
                        "description": "Grados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Grado"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los grados"
                    }
                }
            }
        },
        "/api/horarios/grados/{id}": {
            "get": {
                "tags": [
                    "Grado"
                ],
                "summary": "Obtener un grado por id",
                "description": "Obtener un grado por id",
                "operationId": "obtenerGradoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del grado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Grado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el grado"
                    },
                    "500": {
                        "description": "Error al obtener el grado"
                    }
                }
            }
        },
        "/api/horarios/grados/guardar": {
            "post": {
                "tags": [
                    "Grado"
                ],
                "summary": "Guardar un grado",
                "description": "Guardar un grado",
                "operationId": "guardarGrado",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GradoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grado guardado correctamente"
                    },
                    "400": {
                        "description": "Error al guardar el grado"
                    }
                }
            }
        },
        "/api/horarios/grados/actualizar/{id}": {
            "put": {
                "tags": [
                    "Grado"
                ],
                "summary": "Actualizar un grado",
                "description": "Actualizar un grado",
                "operationId": "actualizarGrado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del grado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GradoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grado actualizado correctamente"
                    },
                    "500": {
                        "description": "Error al actualizar el grado"
                    }
                }
            }
        },
        "/api/horarios/grados/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Grado"
                ],
                "summary": "Eliminar un grado",
                "description": "Eliminar un grado",
                "operationId": "eliminarGrado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del grado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grado eliminado correctamente"
                    },
                    "500": {
                        "description": "Error al eliminar el grado"
                    }
                }
            }
        },
        "/api/horarios/gradoUC": {
            "get": {
                "tags": [
                    "GradoUC"
                ],
                "summary": "Obtener todos los gradosUC",
                "description": "Retorna un array de gradosUC",
                "operationId": "3361b74a7775ecd1b68862c20b619eb6",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "500": {
                        "description": "Error al obtener los gradosUC"
                    }
                }
            }
        },
        "/api/horarios/gradoUC/idGrado/{id}": {
            "get": {
                "tags": [
                    "GradoUC"
                ],
                "summary": "Obtener un gradoUC por ID de grado",
                "description": "Retorna un gradoUC",
                "operationId": "df7eea4931a4396fdb7d0aef3bb19433",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del grado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró el gradoUC"
                    },
                    "500": {
                        "description": "Error al obtener el gradoUC"
                    }
                }
            },
            "delete": {
                "tags": [
                    "GradoUC"
                ],
                "summary": "Eliminar un gradoUC por ID de grado",
                "description": "Eliminar un gradoUC",
                "operationId": "33bd649ba0a04241e13e51d83f1f9461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del grado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró el gradoUC"
                    },
                    "500": {
                        "description": "Error al eliminar el gradoUC"
                    }
                }
            }
        },
        "/api/horarios/gradoUC/idUC/{id}": {
            "get": {
                "tags": [
                    "GradoUC"
                ],
                "summary": "Obtener un gradoUC por ID de UC",
                "description": "Retorna un gradoUC",
                "operationId": "71409768c13345516ded85be23e3eadd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la UC",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró el gradoUC"
                    },
                    "500": {
                        "description": "Error al obtener el gradoUC"
                    }
                }
            },
            "delete": {
                "tags": [
                    "GradoUC"
                ],
                "summary": "Eliminar un gradoUC por ID de UC",
                "description": "Eliminar un gradoUC",
                "operationId": "6fda0027904abfefb9318fc1554e377f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la UC",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró el gradoUC"
                    },
                    "500": {
                        "description": "Error al eliminar el gradoUC"
                    }
                }
            }
        },
        "/api/horarios/gradoUC/guardar": {
            "post": {
                "tags": [
                    "GradoUC"
                ],
                "summary": "Guardar un gradoUC",
                "description": "Guardar un gradoUC",
                "operationId": "03aedc1f8fa4b0fe37a465a236c69ea4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GradoUC"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa"
                    },
                    "500": {
                        "description": "Error al guardar el gradoUC"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GradoDTO": {
                "title": "GradoDTO",
                "description": "GradoDTO model",
                "properties": {
                    "Grado": {
                        "description": "Grado",
                        "type": "string"
                    },
                    "Division": {
                        "description": "Division",
                        "type": "string"
                    },
                    "Detalle": {
                        "description": "Detalle",
                        "type": "string"
                    },
                    "Capacidad": {
                        "description": "Capacidad",
                        "type": "integer"
                    },
                    "carrera_id": {
                        "description": "ID de la carrera",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Grado": {
                "title": "Grado",
                "description": "Grado model",
                "properties": {
                    "id_grado": {
                        "description": "ID del grado",
                        "type": "integer"
                    },
                    "grado": {
                        "description": "Grado",
                        "type": "string"
                    },
                    "division": {
                        "description": "Division",
                        "type": "string"
                    },
                    "detalle": {
                        "description": "Detalle",
                        "type": "string"
                    },
                    "capacidad": {
                        "description": "Capacidad",
                        "type": "integer"
                    },
                    "carrera_id": {
                        "description": "ID de la carrera",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "GradoUC": {
                "title": "GradoUC",
                "description": "GradoUC model",
                "properties": {
                    "id_grado": {
                        "description": "ID del grado",
                        "type": "integer"
                    },
                    "id_uc": {
                        "description": "ID de la unidad curricular",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}